frontend_container/requirements.txt
flask
pymongo
python-dotenv
requests
google-generativeai
grpcio
protobuf
google
boto3